evr: reorg iocsh funcs
Separate implementations from iocsh boilerplate for clarity.
